changed proof message names
