+++ cxs/libcxs/src/utils/constants.rs
@@ -39,9 +39,12 @@ pub const MESSAGE_ANSWERED: &'static str ="MS-106";
 
 // TODO Used in testing, may wish to move/remove these in near future
 pub static PROOF_REQ_JSON: &str = r#"{ "nonce":"123432421212", "name":"proof_req_1", "version":"0.1", "requested_attrs":{"sdf":{"schema_seq_no":15,"name":"state"}},"requested_predicates":{}}"#;
+
 pub static PROOF_JSON: &str = r#"{"proofs":{"claim::e5fec91f-d03d-4513-813c-ab6db5715d55":{"proof":{"primary_proof":{"eq_proof":{"revealed_attrs":{"state":"96473275571522321025213415717206189191162"},"a_prime":"22605045280481376895214546474258256134055560453004805058368015338423404000586901936329279496160366852115900235316791489357953785379851822281248296428005020302405076144264617943389810572564188437603815231794326272302243703078443007359698858400857606408856314183672828086906560155576666631125808137726233827430076624897399072853872527464581329767287002222137559918765406079546649258389065217669558333867707240780369514832185660287640444094973804045885379406641474693993903268791773620198293469768106363470543892730424494655747935463337367735239405840517696064464669905860189004121807576749786474060694597244797343224031","e":"70192089123105616042684481760592174224585053817450673797400202710878562748001698340846985261463026529360990669802293480312441048965520897","v":"1148619141217957986496757711054111791862691178309410923416837802801708689012670430650138736456223586898110113348220116209094530854607083005898964558239710027534227973983322542548800291320747321452329327824406430787211689678096549398458892087551551587767498991043777397791000822007896620414888602588897806008609113730393639807814070738699614969916095861363383223421727858670289337712185089527052065958362840287749622133424503902085247641830693297082507827948006947829401008622239294382186995101394791468192083810475776455445579931271665980788474331866572497866962452476638881287668931141052552771328556458489781734943404258692308937784221642452132005267809852656378394530342203469943982066011466088478895643800295937901139711103301249691253510784029114718919483272055970725860849610885050165709968510696738864528287788491998027072378656038991754015693216663830793243584350961586874315757599094357535856429087122365865868729","m":{"address2":"11774234640096848605908744857306447015748098256395922562149769943967941106193320512788344020652220849708117081570187385467979956319507248530701654682748372348387275979419669108338","city":"4853213962270369118453000522408430296589146124488849630769837449684434138367659379663124155088827069418193027370932024893343033367076071757003149452226758383807126385017161888440","address1":"12970590675851114145396120869959510754345567924518524026685086869487243290925032320159287997675756075512889990901552679591155319959039145119122576164798225386578339739435869622811","zip":"8333721522340131864419931745588776943042067606218561135102011966361165456174036379901390244538991611895455576519950813910672825465382312504250936740379785802177629077591444977329"},"m1":"92853615502250003546205004470333326341901175168428906399291824325990659330595200000112546157141090642053863739870044907457400076448073272490169488870502566172795456430489790324815765612798273406119873266684053517977802902202155082987833343670942161987285661291655743810590661447300059024966135828466539810035","m2":"14442362430453309930284822850357071315613831915865367971974791350454381198894252834180803515368579729220423713315556807632571621646127926114010380486713602821529657583905131582938"},"ge_proofs":[]},"non_revoc_proof":null},"schema_seq_no":15,"issuer_did":"4fUDR9R7fjwELRvH9JT6HH"}},"aggregated_proof":{"c_hash":"68430476900085482958838239880418115228681348197588159723604944078288347793331","c_list":[[179,17,2,242,194,227,92,203,28,32,255,113,112,20,5,243,9,111,220,111,21,210,116,12,167,119,253,181,37,40,143,215,140,42,179,97,75,229,96,94,54,248,206,3,48,14,61,219,160,122,139,227,166,183,37,43,197,200,28,220,217,10,65,42,6,195,124,44,164,65,114,206,51,231,254,156,170,141,21,153,50,251,237,65,147,97,243,17,157,116,213,201,80,119,106,70,88,60,55,36,33,160,135,106,60,212,191,235,116,57,78,177,61,86,44,226,205,100,134,118,93,6,26,58,220,66,232,166,202,62,90,174,231,207,19,239,233,223,70,191,199,100,157,62,139,176,28,184,9,70,116,199,142,237,198,183,12,32,53,84,207,202,77,56,97,177,154,169,223,201,212,163,212,101,184,255,215,167,16,163,136,44,25,123,49,15,229,41,149,133,159,86,106,208,234,73,207,154,194,162,141,63,159,145,94,47,174,51,225,91,243,2,221,202,59,11,212,243,197,208,116,42,242,131,221,137,16,169,203,215,239,78,254,150,42,169,202,132,172,106,179,130,178,130,147,24,173,213,151,251,242,44,54,47,208,223]]},"requested_proof":{"revealed_attrs":{"sdf":["claim::e5fec91f-d03d-4513-813c-ab6db5715d55","UT","96473275571522321025213415717206189191162"]},"unrevealed_attrs":{},"self_attested_attrs":{},"predicates":{}}}"#;
+
 //pub static PROOF_JSON: &str = r#"{"proofs":{"claim::87935efd-769b-4518-965b-a679d53f17fa":{"proof":{"primary_proof":{"eq_proof":{"revealed_attrs":{"name":"1139481716457488690172217916278103335"},"a_prime":"10620565207350874137747328339446227358703052468892818296923658389457577036380653383332623419776704897333596811203766686052873911161412554761694817543901205910282194728764431034873463674958428731435214270883955668698233550631415562295955464775016010535316542208302229677766265208026793749699238109142999625976686929560434063984164145560456005550018037791628467493291850871992428307435950305309742740480943470254421053468950090587009307532681285214213294825823309941110043873461033732095895550172086667159020267656990889092011278845983068743307157244807810360929639634631748276981619801227467121475049627133819732401447","e":"168397498313964963500148577445845944780230231867280406301662326912314760927147263753633388578452409829405601835321430100275565032644952055","v":"989202781258064300391358753006075775312482265107481855635138450734599844089939517372766589807442871020078033866562701706152086704646736653765106040607844796847659767059269714586784858069313744580410516633489719341081512978821133383223014600866845540718747356209963092316463263847351816229401324835921053899283457144251517219236405816608751773000658590004231448099486012346837877066439762141055068032562098387212982409422719268416116363882410510510129674703929434448921828723590338482412966813045160805371070494760127097727100156445729417650922986464211364120138907932150282647314778483845000402911551615320938303542913688512447859309835371056012004207779646764821017477401358008505331045383378744763080578842197081074099499118321690553252944484182529145442058262352655232715168326503879091579339027992897383303289641657668017824691734027503318196244539177797169328057727843428811317787124304882195435989890837622641495899","m":{"age":"13904485317561679035066421555221383686584910629578014436692032488535192188590672351058608089602536812870827863217256556433835982448541652305924699138055185139763732263099283257406","height":"15353895226517514976073647136084623499502612345923675026254384729420729087884766918518359145034983869778926889713067916573697108942512009078468944123602357781333418444139491035201","sex":"4132196222712316800912049331030300081653258064163748471951078455025500801597411450638293446832106288846748980376003401829557519293941044870016686814083965915498814435463213760505"},"m1":"104273093680198956916279670417351737903882886070058277174605894487728727078527282472629105240112599639834187910294030748277118912566626709286962950843437653153783889931702113170747381733487732659815176769256221604475227440322298723932005445665726531014758317417672422558272366259283509629426430792820200084668","m2":"3692283487802029586577429218452425436775884845341721587876057360566902994646046398693223024509189798896249977793107720795908364785683432581985856023851245030999531493910804882963"},"ge_proofs":[{"u":{"2":"16051609818368455146005004727102354847540036918843526796082032627247759155855540151147684469531488477394718370761630815630700762365110661655259342865572759177426945437560293299546","0":"9048232629445944901395850841764419373727734311294466150282838968242841781473546536776144196481832743517627356302850134818088792640030521759329773457157050148035344771550116056158","3":"2825622748327536391861190987974067275810178948068137869341033174003563359072020506491673226777173033052099962622474188057820773299765129299247498925424581186503937334089027329532","1":"11569946736689572389737151999388478147164183857788333079182772215629130067503675016146212269321087353758229521960996001730035620073622401074071382423100676761946814678447953033711"},"r":{"1":"1196718102106704284583477839445757312709454972283811270088019205414132175226384745833930634875170966216150130706901245819702803437991583514137693947770259848457259724405293941014282797662627857857764224549740794331897975781042022264108375392746232656197102355077166521667557512861779992389735380811475146122207500410241667042281999541070793654573793468081986904899289506881040641433354728293164758923368353662131631540689738662629415142496364992242341745570547167740063778988548675807268639896112733807000487920525314094328752655024138116385048681043089222725049075626674215889732361648097668119214577994489072939966617334040738001862014782767594790497648599722439799932039202757166642709821557829712910677705215997029","3":"1116851110618064173499249793243786271032699980859315356307568467913466488554313079861321403237845404408498378189101850795499635477630921560249209622662721591694572732253199859723919659858277165761656762021387504131322465918887466368058785333230621143340358848115523142382425314488055393543725595964128395848337138281443712375721659670235359154295538855607681187545465580863170191380192452688979591938115409499040173361346835345622749696308118035003158976306860356908611926210549118111659193145100040383676224236480357395447429798857432555675145094827917893207155718565661011223188136901574666754760876539404634570559687890143488800986862713352274780372122522956830800887899497967685469992218214863432658457036416238766","DELTA":"37797948278444089654601465859653645744377241364932074451662698811367320593167751087476394670072097745731885070176146144692920268729592037209370742790149918461827431003986912013236536275120607836035027403223327900229433844068935681616596331405644810098845781615338423305388596406021111023359772027241455558824593662170851545016516335216617271142605098108052105153054930917123120537287110459191930083681568058852555316086503456647440959038416083601433176977968406041315567629470831708124473677822540227944117809286106209013800966344135309746698532811634284816334322797436070345500211228569771046465897701825814851319425197849956035008592964731382666165499059419803992798879837006855353751578256259454358098012502805852","2":"1284269377728737031583035169064185162969814675521499270185619262480492201736022763510276649831927872252959097610023329533667410768620805803317219420756008830925814101503241782082917713617823713927506526972327266648229726344304419648708932753792230531560440639030663872902653202129154922926118687758315612007592183635931438115339002971242106300181337501412607778882398455747017019921297270123668759497432710422633590773677115389021637275334403333244971008027923514915157959840100710865193006421249695861447563759622854733896096672235258588848180085472353920350561600869746704971137509003976217708694121307877550633677728666086841909969982892797065537794090431991731823633256239106093062120050350173699980369813431128113","0":"813499694772103216385851218626322236210800573983344799897651964903711133811523304749103809395553046550222219672187587703057995715443972513156055892069249666825819961778054737191682158748448562499343622144337867292916859673213180910174654732982261215187285668863463920803678163073715814660782470045562337664885616011835989179360860670977681869557684326240226623825196910840205817087621655112314190928251821718436425541940066992711776852761679264767265305284171030699040309194936278777056003562593048231166133753609072039917083593643160559537583881726923811457825688547302894585263777864475427872631452698329828736641962386326147183230620386477545287005122853842182829633043330183386850376905197429152047314778099132376"},"mj":"13904485317561679035066421555221383686584910629578014436692032488535192188590672351058608089602536812870827863217256556433835982448541652305924699138055185139763732263099283257406","alpha":"28880093396279840243821945299440752360594719817560663532744147963245444762797528130147090895528142525338777346878729035015116310423223884975229418324581402450119162105535439399054494831794134524751905950323471945375630094935789005941416062312262688042015156389959491836095090314833427861444530149655157001533115291821167419668769748233268259996187329760866701555585904196524409475251643640670220319580954031339843939130120470595907819192630413143237365287340773088508553886108853080147383186568312480743975582304178853944596529139059262527510087674892761589520303174269928801790654365268721875669217953354122561716843580278646381684862227602387007119493218285932311272022015429354449074883232526636797424518174437141738195705872396068674989671453775503999604280394622781655860491295347212872682604094894014553777518823223398246891129243188","t":{"0":"21212105934815902516233234204913736081936826689372063397118655521500851609764704662963727315086815734246315418119804671721341020795196280022077957318336326903434961108385388355874643447821745151425154503785362662286783819801912851531182389276695384215020609517289195189289396566721240730886874878935976933701946994169106547127465199825765917966324902942991328866556054563613126307777438404560686824985427302003259119279047262095979282849964396407748790639881963818893876558721052530830676713901143779116629990058755206949809976491138618039503069216615778574109386977510115959623832293780516628235010550527496540346572","2":"28969053900605379729730114271697815157144465077537743388160064249937479997554080122101153726603038725854899786053029362360201287331549463779214811861592998006969747840329246798518347789581564551282584377546179353284413186804658368300929264077714940773579039914379983094238837356718921978792403051055910946766740972948683007909191436689629974845934353528887399264090844120677467337779624611364592558000725140939554083888775113787834202919060875428501595606053305586776396155314306722607162836930411696738951005261770006779334298066039174964470865939935453121385562671014655143427998549498093153733864684187967207710736","DELTA":"34231194832685737212693958911330612374853613347404215823172274501868528364243141618919918644532359397423273291131252680451018526247575448387724516441683422702852538541810423644545187700615529906904758904218247539367861795606502713321596076068708087606267571527664280636598607019576229368780230510066804669233490762826770436245167331218800699222177354596541511653009648993717692698490701043526065317580333143812548134240593043441409540275970386460762045947979332491633333364981422019120885134304622860911867327965825452635142049076559258806907866813551708943869253104263175119444045030998730362845126559159956767433383","1":"61017955603583424531254860538951332414875875565575818042730252033310892080560734656223248453209319811928791467354417586519174834787494299768354866192774528259946487319554915996813085020118788518141788949877100053056979996343323017237543671385980182375136504248420719578620267062390932336012970997126864706724262035305891236244180181268750779931088866513188122394200649455003644432596066060814988335925981012707730974403956234226869790507862814427403341155648748906479108345054837699801558798262982421669155226040790794292259831935726466786955005178086059875285402387708751048637331533605697580102694565705207700896138","3":"80279607567979104043373457545578667434142623124550991477170757042903593150799324362547515110266844809502087767127130506279924661273183637016087288841739311524878872093177172505367066498884405318139316416481125871669117580623750738831432081771273061600182917934801605042839711908088773495181175728211373751346366293188469905870255153488284732631021303909336760423104843130723303599587373480453760181707875948294300358373534895575665270075278411316653391151213650880090074648412452586575493915229524668971498424344910321363068517076097443918478407504602120938321459500030135129820153252874069799157520914109086651034160"},"predicate":{"attr_name":"address1","p_type":"GE","value":18,"schema_seq_no":null,"issuer_did":null}}]},"non_revoc_proof":null},"schema_seq_no":15,"issuer_did":"4fUDR9R7fjwELRvH9JT6HH"}},"aggregated_proof":{"c_hash":"38688359780520913226176025726519080004522675231256989274715137785775498982984","c_list":[[84,33,142,153,199,183,174,203,53,85,180,165,83,134,152,198,83,40,168,215,184,122,208,251,19,207,174,37,234,190,154,150,189,46,155,101,179,234,104,104,148,20,158,73,200,7,64,158,30,207,65,221,112,74,128,77,56,221,39,61,10,28,210,152,6,247,82,33,206,98,152,217,95,5,226,59,12,135,78,145,186,5,184,1,157,207,235,104,0,91,110,10,91,96,197,68,112,13,209,218,68,219,145,124,65,131,12,27,177,109,96,226,122,13,232,131,11,51,127,78,234,69,170,24,236,239,198,50,202,141,236,61,126,109,219,161,240,13,180,30,124,143,91,59,158,28,155,193,240,228,196,13,206,37,184,59,59,84,114,72,58,163,227,17,151,144,66,244,221,94,89,120,31,23,95,121,165,53,131,31,253,155,139,239,235,235,78,99,141,150,25,62,93,235,117,171,56,104,70,215,6,104,228,69,66,172,43,93,74,101,191,213,116,129,142,128,188,204,91,30,184,160,180,111,72,184,223,209,63,66,172,191,105,55,156,184,53,138,152,253,68,167,6,41,135,164,74,191,131,124,110,7,21,204,157,39],[168,8,65,72,200,131,6,234,232,142,153,73,228,128,67,21,2,113,192,207,145,234,10,251,170,197,61,62,237,90,109,251,220,28,80,171,153,93,82,71,159,147,135,21,89,255,220,4,98,33,56,117,144,229,142,45,186,44,119,143,242,54,196,88,202,166,75,112,233,163,7,108,104,129,48,221,153,248,206,219,235,62,166,202,239,46,134,35,124,240,102,252,57,8,242,61,120,37,199,186,28,248,178,39,40,41,47,103,216,21,178,115,0,254,222,80,113,230,64,80,151,145,79,147,53,217,156,124,192,22,76,209,83,74,79,81,21,27,154,167,21,56,172,152,37,30,252,139,55,76,142,133,217,143,45,147,48,42,201,164,113,47,75,155,178,39,30,219,228,158,65,55,233,6,144,220,72,125,21,3,15,157,156,231,140,28,119,38,10,178,81,52,182,137,36,123,251,13,99,201,81,5,236,32,89,90,40,251,4,42,195,249,68,156,94,41,213,14,199,111,91,35,175,99,35,215,105,254,148,14,84,28,223,77,39,10,203,68,109,190,52,253,165,95,155,102,122,186,155,68,64,222,217,39,88,204],[1,227,90,245,228,191,125,59,180,143,126,254,190,29,192,160,147,157,126,25,130,169,124,149,250,201,25,30,207,75,243,60,228,205,113,119,103,244,173,205,78,37,149,55,37,165,75,217,124,234,60,220,241,203,52,159,82,235,95,125,57,138,99,11,222,107,182,138,215,35,138,113,22,191,37,44,193,193,227,97,106,45,115,39,72,57,206,241,128,89,109,215,213,38,143,251,9,122,108,137,14,125,251,250,73,250,255,230,99,111,65,169,72,170,169,231,211,255,208,50,99,95,108,249,203,64,76,138,212,61,20,198,202,19,11,201,170,60,248,174,61,197,110,128,43,18,169,72,35,61,2,159,167,88,236,118,46,253,183,84,129,80,93,20,64,1,103,7,234,203,10,167,214,56,17,41,75,61,212,184,158,155,5,9,81,218,172,195,6,220,16,98,92,37,190,54,126,207,150,162,253,42,195,246,93,189,89,140,80,100,90,131,167,150,206,63,140,99,182,75,7,28,219,64,195,199,75,86,71,21,135,83,179,180,188,76,145,198,115,208,175,84,202,22,67,52,119,142,66,190,219,150,145,201,113,185,138],[229,122,166,247,210,194,98,228,63,124,28,111,79,238,133,120,52,238,53,248,243,16,180,236,50,164,28,8,102,11,3,189,157,145,108,52,19,201,121,40,240,59,47,46,65,66,21,166,94,31,229,201,254,42,65,17,71,45,197,126,175,170,249,18,59,80,232,20,24,252,143,143,20,177,163,39,184,109,206,120,236,120,217,1,209,218,231,206,189,221,156,63,54,114,63,7,75,162,91,5,162,66,60,105,100,64,227,19,217,58,158,16,59,88,7,252,93,109,38,222,14,15,60,7,10,168,77,25,45,36,84,89,193,20,107,171,7,1,198,58,238,206,98,170,199,14,28,255,12,17,148,200,163,217,60,208,21,217,202,210,101,172,129,98,103,204,115,55,163,95,183,152,109,162,194,47,150,88,242,22,152,105,206,117,14,183,239,53,26,74,32,203,51,224,97,11,20,14,136,172,84,199,241,7,21,231,117,58,233,25,245,93,248,162,49,193,24,110,102,11,220,169,50,160,221,162,159,36,68,206,77,207,97,164,43,204,114,25,186,253,216,251,194,210,9,241,66,125,60,29,33,96,43,163,32,16],[2,123,239,222,132,91,146,91,110,85,232,250,48,213,162,247,125,233,128,181,226,227,228,137,32,95,7,198,50,219,4,197,224,127,52,46,202,49,238,218,154,82,29,159,184,242,51,199,20,206,162,80,72,184,192,95,200,225,126,248,173,209,224,92,188,152,1,198,169,24,131,150,46,166,98,252,42,23,187,55,245,132,54,166,167,240,57,253,130,1,191,82,15,13,36,114,143,213,62,145,132,205,46,61,88,190,48,18,181,143,240,124,183,185,191,72,149,165,110,82,226,69,236,9,37,68,30,247,175,249,180,59,77,73,242,90,59,6,57,13,220,142,39,156,172,157,11,112,72,59,0,175,126,251,189,88,255,171,224,223,28,19,235,92,102,231,67,211,36,207,126,195,58,164,155,209,1,122,48,94,152,163,181,38,252,250,160,166,93,142,31,103,221,48,55,152,7,243,205,196,204,67,14,134,200,253,206,199,232,135,94,238,29,143,135,229,112,161,52,206,233,159,192,135,167,140,84,203,109,117,62,177,193,100,147,247,91,83,206,119,198,207,239,124,152,114,58,137,238,255,87,176,223,22,10,170,48],[1,15,41,205,193,40,170,76,71,41,247,41,56,4,13,235,82,176,108,21,152,150,203,57,46,30,183,210,178,2,184,41,196,20,114,150,72,44,221,106,111,181,70,241,124,193,241,217,238,161,56,197,136,127,9,35,71,45,1,78,248,150,179,187,102,44,112,251,235,122,172,242,69,109,86,236,145,233,255,27,60,28,166,139,76,41,137,156,143,22,103,170,45,146,52,130,77,146,193,54,201,63,101,144,165,2,122,109,143,182,160,93,107,2,239,226,83,31,199,83,239,51,30,230,208,224,7,7,126,92,21,197,207,75,149,78,87,56,171,232,149,196,48,128,108,24,208,111,119,165,175,169,177,34,232,38,6,201,116,145,63,255,71,232,24,61,82,41,164,128,137,218,5,104,92,226,134,155,5,62,47,12,184,43,144,79,45,95,49,7,77,204,100,132,74,145,218,19,1,48,183,127,140,11,67,108,32,227,90,197,64,106,222,194,254,22,63,60,250,58,215,206,124,75,99,220,6,84,72,146,164,227,26,143,170,128,214,126,188,143,221,158,125,116,69,112,41,79,118,208,211,3,117,151,22,226,167]]},"requested_proof":{"revealed_attrs":{"attr1_uuid":["claim::87935efd-769b-4518-965b-a679d53f17fa","Alex","1139481716457488690172217916278103335"]},"unrevealed_attrs":{},"self_attested_attrs":{},"predicates":{"predicate1_uuid":"claim::87935efd-769b-4518-965b-a679d53f17fa"}}}"#;
 pub static SCHEMAS_JSON: &str = r#"{"claim::e5fec91f-d03d-4513-813c-ab6db5715d55":{"seqNo":15,"data":{"name":"gvt","version":"1.0","attr_names":["address1","address2","zip","city","state"]}}}"#;
+
 //pub static CLAIM_DEFS_JSON: &str = r#"{"claim::87935efd-769b-4518-965b-a679d53f17fa":{"ref":1,"origin":"NcYxiDXkpYi6ov5FcYDi1e","signature_type":"CL","data":{"primary":{"n":"98323539446717710543395662336566658660102799649633398701532674094104742695061842030528886102886608126390483575252846240674386327190977261903819256001621732326488758477154749318185651960441563934096552665993420885198459189297542176502716706456475927859439592743426562482380774181922664996519722378981881679233920750255713448843436437866259186576548077842754373602760889398604316632776678455666819962898884624397156670589122843446607123110954008416297624245783528343373339330999445869728108075077622302555609629649509117306691844208966464911750727940083294531175143673985573204005493286194901325160736039872094634054777","s":"77988317946494796977065389806431538314565940574588040032683561085447422552921156875471033593871583917842576467069041598653151473163976829661417754054148545336212726250227765326140467116092933229114303163102648642111578144530022004784918886832223251321823775468223164229301142475421196677889316488627288928606962259878826618845943157825371924905679508790192203000405771057999799697018678311751059133769020885771204924600651273015127343280795209010946969733571653464563618882066855772659402701952437620778856301891086499020968436926672448095363273997101651962476881779584399013721288870116490530264443199188013216732934","rms":"5644026678492094278524225514523722954666324657954561086499528452403522776045396376929619583065755173101472533739008564352112782371929580410855776978724132849647818691067944375140537565987661696790336454931029539911347033585152360733143357514895292119500556557145123836791332883944812220117207048078498143183389958166646333272579411254435050384530401302181046043084620109608272239153388875992864882193815192196705724868886110670808182755658638007370336237513441857345367155275977910987474011916315618532477549345117506081574296984414394063013756066866090324856784857922100744411587454151980198229704131537076948646478","r":{"height":"74367497409787115429076751486860890776225661379106167520735117439024955550792507195126272018480668737896179995724242077094851355673466392095373280379988686969360945254827613486335146757338092989523896292008787012056008847951385054048394388866945284745299393817006579964702752344957783575975385917673626039009883921230089892899838193451319636869850968876502173667334435795350595826944718339987371086871204928233430916334112797644557047199905381234727884543202551847547363102579987095546728981623694458358748717326275478440592580313093530977268817136738631066319304804014540109795388571769220057370143749644475648167710","sex":"73600112020860694988070697471473250227904303990603016584429039206604147237079273023649935727381975224105750260191125659159799454811611221209403365200480259499450905468077829096643095842490388898319083164963680249798069467438572454803211226814075995806679149058135737546562473739586253869214146691476478766464980578915521912250772026906660192620914726974429865832818630541546046801314756298018039084217173580129753281132198897918968429772403420046673684511659032273066215174810614668558730414129960778706979142769543064529758089266837256771082803878056758840184700794663165005414336137782360166841274813548771581560589","age":"47772732570272498719193575185972017499522134812966255940441488143115950763651054544662093686367240147773454886139427668604891949504818154877211512352352758199611329744841936015888101816545902250935993071626392551640030508879957517377864981464879154099680111062216317914713063795472313676120296695845006476092053896326658331843759315593741772862566360818156421374531983374114022081829302382877277122024553272260679458251025781427310567283208688622141450675697164814556695729167516271176987554948791059358479599626329623227908669956227710207222077393804079934807386198899701229885757573709690147294837099240682042327691","name":"4969629918641657003165912163546219309925162228252419241346207684394995737903340882741555108092981297394869196204180372195280466354844835687769451371798860074357397551954285676210923931012068112419310219571768502455902949132287299683518604490586768182062290206878965726888858478236889360222302022855431034476502762004203498084813245221527171023783571857581652026515110019089148903298558205785208497966967526342328686910574681592677375110723677950197040847047966357755583363452373538086601182020403348955921662527114969223180852244038099676446458304724386662596841197709488044401669841785406763371643801299479069785013"},"rctxt":"29111609892008843232896399527704335884071266627824219856513144868919188746668251245074846045811585875883458746747418347105410571398363792714915908705885821119214880031000040221308972233271160515941401401798113070754522326685845494052374541269085458187816953404850332588758307108015031183593355358161607672955377208479919896919267738186641548953142680946815722172177889883574106000555043621897995396005248657512595536280932853036671448511076347512139359723986839471402373140402122167041259834826399477138984862421376805368412867992526194566917748684101018951370043228863989598392874077268060549214871132362607900743643","z":"66249945907746222155601751142256558437314066560951012450688920210246347321792375982645176650098659239499242177848304381610690614741728928060402527127318199681020663253992190269982763600004943891271551042193468689058604482115922104139640106183048181095382621106714563780180051567181476431925942012053022532548776578179647329453686428219067363097515684053915682173723581641141232162523650167386463654713970420934288475211936320757622175606421923752608863940402422011335194061737781342797178787433053899666777931840040600296471824203840504329544264136130911239907036144182076199393267171596431953760984579482085269620670"},"revocation":null}}}"#;
 
 pub static CLAIM_DEFS_JSON: &str = "{\"claim::e5fec91f-d03d-4513-813c-ab6db5715d55\":{\"data\":{\"primary\":{\"n\":\"104050607798814202443347751990505605058079821103993809294351961155626604457119457502254315147872263788175111883696626719477687687851630201683172254223814733078876255145394173960074484234978164795838568222532729171888794536652794334457694250665609531060339910889250562443037434097815004083713682128773364697233010473026651125795070493957354393804400356149329922474300332714868344029297112981585225154394477549127626669985189159660176093553633698062205822712936786303940694357718548007109219277207199047379357176793754168776500505959654325698856068009139437969448400481847783307374883434582612029048201180092511407001901\",\"r\":{\"address1\":\"14627094764153421576277267435752420018656890275418956183490011204214201436779681040772687490602158626262546809022555212118752401016629202044045632036686818475977498601831376398611623968912025014246774572815484255678230621640829565615723321104762752531283660596274198858275340738764166743434480281531386669786337905643854049798378608726889603623224387030399667274838106295506860774549084747712076686184406070751119294242658512224718690073936408659829736803160102780455301180754781328846424565610712246053765107774374487434264034314714485553756851923371984656303296606739253431778207619039113172546182637783056733552398\",\"address2\":\"92660006765065878540242856121679957310156890016625265597674496021630294220787391658907268773120722800050620756162683338209066199098670961498097562900525345770867006984558198583451689810968196474707836261787174820019004913811455697044210492019739440427563973141408212819543569105096383195331578788660623397794776239823253025688481706368268253259270514070675908119766099141207670345699281398701116144318939444295662169101532679056485812349629143420797527369552592124744966897032136715932763466461582037011726170575571845296707603328738766693461701560396853446209744442937934269716424935024546492328856204750558888851371\",\"city\":\"56075795090313195827197096946600960086129603693920569063858703968954904982889220480613892672877932729824301888088959078704092691697351095066369674866054724955421147206990787122623542166706866269729427390611965035454965866683019873899663217139102382356537360132504210260373758796134386800397115233372456491768176278363610670604340693349050420757865173243398625711860080511431174926305685843349086478965443559993011199177441146498217449981478217964981453112226323575536657289803983166266587292903873151025151826448237247800311400569960717520890819512464694197017039627502014413036467544798585898990871989307067743852938\",\"state\":\"63481090662541157570539528741613894480850534885129939135827419986820268839944049517606446847948481260844413569630064814347982673978232471514387231785922245414908446805288181013170263865193207891794670053898437362061514649907756611389994339655257537052079866800607353048076236237107796459292146933582167208296799954757880598035301604705234765944181148146916034847511285659180569373890532057062979827198978538589754034576850593136627137330713890438771662595603945904978959970615252628681049471483819946926450457718825342401872694497619152478359070640920307715225636393967209150130481920206364305567453335471348125357222\",\"zip\":\"57234405006618584645765087416477224271826047583224381623166567451987903857927057577338953775503496864249507162212516286951770303123780647449775974877093114650145482620636033103828199699585795106777748684823616382081536055148547705428833321228605365979545397614414702703721344013175754282805364860677584741428547483495201197875925409456178672940376998246247970501748224647047682828278574498579750300047165905940345160301141654783739470632945631308877254732452563813851923970793493514286658179442107053311483835773112529011357639927776280873065839513548435718696985936833241898909806443016083284757281623583734676593409\"},\"rctxt\":\"53554802941693967101317526248692505949051527802253741687364450544763543887104669901115344280362982526851996639310839032481377687143311032134640211094646299647570216363801078938470843736815456872292531328764640927687660807040816619635193845745795944612876615961519041429783999936121084751656927253295075981502168644235927329397601842537723963994050434753854947731230957441863816482259778997424358606117340085124972741191171351000201346784841882602097577974026697961476971604751239913353339159079158150172743215153389390276608517590140962592561002560609618606894049045145580356690212302688484119285040074658899883460125\",\"rms\":\"1405245892288625349510224954068867138414982694432904322559372329796476735644544135423656292520298416990204791157932377278529888484558400074982573316444942524482475459658338843713743589372443156042347734945950766832177951900497466813488717812341831925053316636453026522587258508587341565788823419202590643449746689692886761696269299920434418734593309688930209908205972147648466709733505697943979190375743768640344334436711847434477833491428785025465969442264987901998925926625368207149997641658874305875458809379572966198742764316678000211113107162061800709052398019084533638904584038079068153812377744872128630988429\",\"s\":\"19914031557821425761560831920477871184070459895910700572840755915015498762816033608860369225901996598888585867724025139931729651216232721475043725563711229010974655176231790234973380823326976111121099086553666816452484781955644128329589244366179245071574783462301365456067725499844966617752271832826210569703857001268664207657429716035223213266406424997544247414143668237294983143730821696515854435773406075096948291591901696871187866186780319889658857479093023482498745793935240694400533072062401178358980716400074200999056971176166951081242840624277049268584579218457644680499991424397178039371485163564173114086249\",\"z\":\"20538704467177514098291269088542581021966379009288778802665868131871112391324748167652613996162973556529343040162803411371221183462142196727263656167188965436466021158636226533528848880270622491320889545873259587098543607417791618396592663655869076718191329644155926732886216836709036448217179896342505107616205996809773735923245687673822906326539820896470952322615173526878951179530922918658736789945002113278973732242839100444613969402461463600179541900155332916169262711371228352032443273469311250772911494092822677840217379028600657884454842484171620672228234276833120319300163863846510541083130200675548350591646\"},\"revocation\":null},\"identifier\":\"GGBDg1j8bsKmr4h5T9XqYf\",\"origin\":\"4fUDR9R7fjwELRvH9JT6HH\",\"ref\":15,\"reqId\":1513355887705828001,\"seqNo\":20,\"signature_type\":\"CL\",\"txnTime\":1510335473,\"type\":\"108\"}}";
