+++ cxs/libcxs/tests/demo.rs
@@ -17,8 +17,11 @@ use std::thread;
 use std::time::Duration;
 use std::ffi::CString;
 use cxs::api;
+use cxs::utils::pool;
 use std::sync::mpsc::channel;
 use cxs::utils::wallet::get_wallet_handle;
+use std::path::Path;
+use cxs::utils::error;
 
 #[allow(dead_code)]
 static SERIALIZED_CONNECTION: &str = r#"{"source_id":"test_cxs_connection_connect","handle":2608616713,"pw_did":"62LeFLkN9ZeCr32j73PUyD","pw_verkey":"3jnnnL65mTW786LaTJSwEKENEMwmMowuJTYmVho23qNU","did_endpoint":"","state":4,"uuid":"","endpoint":"","invite_detail":{"e":"34.210.228.152:80","rid":"6oHwpBN","sakdp":"key","sn":"enterprise","sD":"62LeFLkN9ZeCr32j73PUyD","lu":"https://s19.postimg.org/ykyz4x8jn/evernym.png","sVk":"3jnnnL65mTW786LaTJSwEKENEMwmMowuJTYmVho23qNU","tn":"there"}}"#;
@@ -30,7 +33,30 @@ static CLAIM_DEF_ISSUER_DID: &str = "V4SGRU86Z58d6TV7PBUe6f";
 static CLAIM_DEF_SCHEMA_SEQ_NUM: u32 = 103;
 
 
+fn sandbox_pool_setup() {
+    let node_txns = vec![
+        r#"{"data":{"alias":"Node1","blskey":"4N8aUNHSgjQVgkpm8nhNEfDf6txHznoYREg9kirmJrkivgL4oSEimFF6nsQ6M41QvhM2Z33nves5vfSn9n1UwNFJBYtWVnHYMATn76vLuL3zU88KyeAYcHfsih3He6UHcXDxcaecHVz6jhCYz1P2UZn2bDVruL5wXpehgBfBaLKm3Ba","client_ip":"34.212.206.9","client_port":9702,"node_ip":"34.212.206.9","node_port":9701,"services":["VALIDATOR"]},"dest":"Gw6pDLhcBcoQesN72qfotTgFa7cbuqZpkX3Xo6pLhPhv","identifier":"Th7MpTaRZVRYnPiabds81Y","txnId":"fea82e10e894419fe2bea7d96296a6d46f50f93f9eeda954ec461b2ed2950b62","type":"0"}"#,
+        r#"{"data":{"alias":"Node2","blskey":"37rAPpXVoxzKhz7d9gkUe52XuXryuLXoM6P6LbWDB7LSbG62Lsb33sfG7zqS8TK1MXwuCHj1FKNzVpsnafmqLG1vXN88rt38mNFs9TENzm4QHdBzsvCuoBnPH7rpYYDo9DZNJePaDvRvqJKByCabubJz3XXKbEeshzpz4Ma5QYpJqjk","client_ip":"34.212.206.9","client_port":9704,"node_ip":"34.212.206.9","node_port":9703,"services":["VALIDATOR"]},"dest":"8ECVSk179mjsjKRLWiQtssMLgp6EPhWXtaYyStWPSGAb","identifier":"EbP4aYNeTHL6q385GuVpRV","txnId":"1ac8aece2a18ced660fef8694b61aac3af08ba875ce3026a160acbc3a3af35fc","type":"0"}"#,
+        r#"{"data":{"alias":"Node3","blskey":"3WFpdbg7C5cnLYZwFZevJqhubkFALBfCBBok15GdrKMUhUjGsk3jV6QKj6MZgEubF7oqCafxNdkm7eswgA4sdKTRc82tLGzZBd6vNqU8dupzup6uYUf32KTHTPQbuUM8Yk4QFXjEf2Usu2TJcNkdgpyeUSX42u5LqdDDpNSWUK5deC5","client_ip":"34.212.206.9","client_port":9706,"node_ip":"34.212.206.9","node_port":9705,"services":["VALIDATOR"]},"dest":"DKVxG2fXXTU8yT5N7hGEbXB3dfdAnYv1JczDUHpmDxya","identifier":"4cU41vWW82ArfxJxHkzXPG","txnId":"7e9f355dffa78ed24668f0e0e369fd8c224076571c51e2ea8be5f26479edebe4","type":"0"}"#,
+        r#"{"data":{"alias":"Node4","blskey":"2zN3bHM1m4rLz54MJHYSwvqzPchYp8jkHswveCLAEJVcX6Mm1wHQD1SkPYMzUDTZvWvhuE6VNAkK3KxVeEmsanSmvjVkReDeBEMxeDaayjcZjFGPydyey1qxBHmTvAnBKoPydvuTAqx5f7YNNRAdeLmUi99gERUU7TD8KfAa6MpQ9bw","client_ip":"34.212.206.9","client_port":9708,"node_ip":"34.212.206.9","node_port":9707,"services":["VALIDATOR"]},"dest":"4PS3EDQ3dW1tci1Bp6543CfuuebjFrg36kLAUcskGfaA","identifier":"TWwCRQRZ2ZHMJFn9TzLp7W","txnId":"aa5e817d7cc626170eca175822029339a444eb0ee8f0bd20d3b0b76e566fb008","type":"0"}"#];
+    let pool_name = "PoolForDemo";
+    let config_string = format!("{{\"genesis_txn\":\"/tmp/{}.txn\"}}", &pool_name);
+    let nodes_count = 4;
+    let pool_name = "PoolForDemo";
+    let txn_file_data = node_txns[0..(nodes_count as usize)].join("\n");
+    let txn_file_path = "/tmp/PoolForDemo.txn";
+    pool::create_genesis_txn_file(&pool_name, &txn_file_data, Some(Path::new(txn_file_path)));
+    assert_eq!(pool::pool_config_json(Path::new(txn_file_path)),config_string);
+    assert_eq!(pool::create_pool_ledger_config(&pool_name, Some(Path::new(&txn_file_path))),Ok(error::SUCCESS.code_num));
 
+}
+
+pub fn open_sandbox_pool() -> u32 {
+    let pool_name = "PoolForDemo";
+    sandbox_pool_setup();
+    let config = r#"{"refresh_on_open": true}"#;
+    pool::open_pool_ledger(&pool_name, Some(config)).unwrap()
+}
 
 #[ignore]
 #[test]
@@ -63,15 +89,24 @@ fn test_demo(){
     assert_eq!(r,0);
     thread::sleep(Duration::from_secs(1));
 
+    // Creating a Trustee DID -> sufficient permissions to create ClaimDef
+
+
     // Create Claim Offer ***************************************************************
-    let source_id = "Claim For Driver's License";
-    let claim_name = "Driver's License";
+    let source_id = "Name and Sex";
+    let claim_name = "Name and Sex";
     let claim_data:serde_json::Value = serde_json::from_str(CLAIM_DATA).unwrap(); // this format will make it easier to modify in the futre
     let ledger_issuer_did = "V4SGRU86Z58d6TV7PBUe6f";
     let ledger_schema_seq_num = 103;
     let (err, claim_handle) = create_claim_offer(claim_name, source_id, claim_data, ledger_issuer_did, ledger_schema_seq_num);
     assert_eq!(err, 0);
     assert!(claim_handle>0);
+
+    // Create Proof **************************************************************
+    let requested_attrs = "[{\"schema_seq_no\":103,\"name\":\"name\",\"issuer_did\":\"V4SGRU86Z58d6TV7PBUe6f\"},{\"schema_seq_no\":103,\"name\":\"sex\",\"issuer_did\":\"V4SGRU86Z58d6TV7PBUe6f\"}]";
+    let (err, proof_handle) = create_proof_request(source_id, requested_attrs);
+    assert_eq!(err, 0);
+    assert!(proof_handle>0);
 
     // Create Connection **************************************************************
     let (sender, receiver) = channel();
@@ -101,7 +136,7 @@ fn test_demo(){
 //    let lphone_number = "8017900625";
     let rc = api::connection::cxs_connection_connect(command_handle,
                                                      connection_handle,
-                                                     CString::new("{\"phone\":\"3852322527\"}").unwrap().into_raw(),cb);
+                                                     CString::new("{\"phone\":\"8017900625\"}").unwrap().into_raw(),cb);
     assert_eq!(rc, 0);
     let err = receiver.recv_timeout(utils::timeout::TimeoutUtils::long_timeout()).unwrap();
     assert_eq!(err,0);
@@ -135,7 +170,9 @@ fn test_demo(){
     let err = serialize_cxs_object(claim_handle, serialize_claim_fn);
     assert_eq!(err,0);
 
-    receive_request_send_claim(connection_handle,claim_handle)
+    receive_request_send_claim(connection_handle,claim_handle);
+
+    send_proof_request_and_receive_proof(connection_handle, proof_handle);
 }
 
 fn receive_request_send_claim(connection_handle: u32, claim_handle:u32){
@@ -153,11 +190,29 @@ fn receive_request_send_claim(connection_handle: u32, claim_handle:u32){
     assert_eq!(err, 0);
 }
 
+fn send_proof_request_and_receive_proof(connection_handle: u32, proof_handle:u32){
+    let state = wait_for_updated_state(proof_handle, 1, api::proof::cxs_proof_update_state);
+    let target_state = 4;
+    let target_proof_state = 1;
+    open_sandbox_pool();
+    // Send Proof Request *************************************************************
+    let err = utils::demo::send_proof_request(proof_handle, connection_handle);
+
+    let state = wait_for_updated_state(proof_handle, target_state, api::proof::cxs_proof_update_state);
+
+    assert_eq!(state, target_state);
+
+    // Receive Proof
+    let err = utils::demo::get_proof(proof_handle, connection_handle);
+    assert_eq!(err, 0);
+}
+
 fn insert_claim_def(){
     // init the wallet
     let claim_def_issuer_did= CLAIM_DEF_ISSUER_DID;
     let schema_string = claim_def_wallet::create_default_schema(CLAIM_DEF_SCHEMA_SEQ_NUM);
-    claim_def_wallet::put_claim_def_in_wallet(get_wallet_handle(), claim_def_issuer_did, &schema_string); }
+    claim_def_wallet::put_claim_def_in_wallet(get_wallet_handle(), claim_def_issuer_did, &schema_string);
+}
 
 #[allow(dead_code)]
 fn init_sdk(){
